fix: deletion of composer.lock file, fixes #536

+++ internal/verifier/rector.go
@@ -24,26 +24,37 @@ func (r Rector) Fix(ctx context.Context, config ToolConfig) error {
 		return nil
 	}
 
-	composerJSONPath := path.Join(config.RootDir, "composer.json")
+	_, err := os.Stat(path.Join(config.RootDir, "vendor"))
+	vendorExists := !os.IsNotExist(err)
+
 	var backupData []byte
-	if _, err := os.Stat(composerJSONPath); err == nil {
-		backupData, err = os.ReadFile(composerJSONPath)
-		if err != nil {
-			return fmt.Errorf("failed to backup composer.json: %w", err)
-		}
-	}
+	var composerLockData []byte
 
-	// Check and remove existing vendor/composer.lock
-	vendorPath := path.Join(config.RootDir, "vendor")
-	composerLockPath := path.Join(config.RootDir, "composer.lock")
+	if !vendorExists {
+		composerJSONPath := path.Join(config.RootDir, "composer.json")
 
-	if _, err := os.Stat(vendorPath); err == nil {
-		if err := os.RemoveAll(vendorPath); err != nil {
-			return err
+		if _, err := os.Stat(composerJSONPath); err == nil {
+			backupData, err = os.ReadFile(composerJSONPath)
+			if err != nil {
+				return fmt.Errorf("failed to backup composer.json: %w", err)
+			}
 		}
-	}
-	if _, err := os.Stat(composerLockPath); err == nil {
-		if err := os.Remove(composerLockPath); err != nil {
+
+		composerLockPath := path.Join(config.RootDir, "composer.lock")
+		if _, err := os.Stat(composerLockPath); err == nil {
+			composerLockData, err = os.ReadFile(composerLockPath)
+			if err != nil {
+				return fmt.Errorf("failed to backup composer.lock: %w", err)
+			}
+		}
+
+		if _, err := os.Stat(composerLockPath); err == nil {
+			if err := os.Remove(composerLockPath); err != nil {
+				return err
+			}
+		}
+
+		if err := installComposerDeps(config.RootDir, "highest"); err != nil {
 			return err
 		}
 	}
@@ -56,10 +67,6 @@ func (r Rector) Fix(ctx context.Context, config ToolConfig) error {
 		rectorConfigFile = path.Join(config.ToolDirectory, "php", "vendor", "frosh", "shopware-rector", "config", fmt.Sprintf("shopware-%s.0.php", config.MinShopwareVersion[0:3]))
 	}
 
-	if err := installComposerDeps(config.RootDir, "highest"); err != nil {
-		return err
-	}
-
 	for _, sourceDirectory := range config.SourceDirectories {
 		rector := exec.CommandContext(ctx, "php", "-dmemory_limit=2G", path.Join(config.ToolDirectory, "php", "vendor", "bin", "rector"), "process", "--config", rectorConfigFile, "--autoload-file", path.Join("vendor", "autoload.php"), sourceDirectory)
 		rector.Dir = config.RootDir
@@ -69,18 +76,17 @@ func (r Rector) Fix(ctx context.Context, config ToolConfig) error {
 		fmt.Print(string(log))
 	}
 
-	// Cleanup after execution
-	if err := os.RemoveAll(vendorPath); err != nil {
-		return err
-	}
-	if err := os.Remove(composerLockPath); err != nil {
-		return err
-	}
+	if !vendorExists {
+		if backupData != nil {
+			if err := os.WriteFile(path.Join(config.RootDir, "composer.json"), backupData, 0o644); err != nil {
+				return fmt.Errorf("failed to restore composer.json: %w", err)
+			}
+		}
 
-	// Restore composer.json
-	if backupData != nil {
-		if err := os.WriteFile(composerJSONPath, backupData, 0o644); err != nil {
-			return fmt.Errorf("failed to restore composer.json: %w", err)
+		if composerLockData != nil {
+			if err := os.WriteFile(path.Join(config.RootDir, "composer.lock"), composerLockData, 0o644); err != nil {
+				return fmt.Errorf("failed to restore composer.lock: %w", err)
+			}
 		}
 	}
